+++ hifive1/CHANGELOG.md
@@ -8,7 +8,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 
 - Update `e310x-hal` dependency and adapt code
-- Add async feature flag to enable embedded-hal-async digital module support
+- Add async feature flag to enable embedded-hal-async modules support
+- Add embassy feature flag to enable embassy-time-driver module support
 
 ## [v0.13.0] - 2024-12-10
 
@@ -21,11 +22,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Use inline assembly instead of binary blobs for flash
 
 ## [v0.12.0] - 2023-03-28
+
 - Update e310x-hal to v0.11 with new svd2rust generated code
 
 ## [v0.11.0] - 2023-03-03
 
 ### Changed
+
 - Updated riscv dependency to v0.10 with interrupt/critical section changes
 
 ## [v0.10.0] - 2021-07-15